My database addions
IDs are not auto generating
Check that all array lists are actually copying, not just x = y
User terminating a group association is missing, any others?

Not ALL the database functions are implemented, but enough are to make expanding them easy. I can only work with a single person, group, etc. because the id's don't generate and I had to hard code the ids for testing. One or 2 testing functions were also added, they can be removed.

This is a netbeans project. I couldn't use the pom stuff, so I just imported a library.

See this url for the database stuff. It has everything:
http://www.objectdb.com/java/jpa/persistence/overview